% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAbund.R
\name{getAbund}
\alias{getAbund}
\title{Get abundances}
\usage{
getAbund(x, tstep = NULL)
}
\arguments{
\item{x}{the output from either \code{untbSim} or \code{compSim}}

\item{tstep}{integer giving the specific timestep(s) requested; if \code{NULL}
(the default) then all timesteps are used}
}
\value{
a data.frame with three columns: \code{tstep} gives the timestep,
\code{spID} gives the species ID, \code{abund} gives the abundance of that
species in that timestep. Note: this is a tidy-style data.frame
}
\description{
Extract a data.frame of species abundances across timesteps from
the output of simulation functions
}
