% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{untbSim}
\alias{untbSim}
\title{Unified Neutral Theory of Biodiversity Simulation}
\usage{
untbSim(
  J,
  Sm,
  Jm,
  nu,
  m,
  niter,
  niterTimestep = floor(niter/10),
  initLocal = NULL,
  initMeta = NULL
)
}
\arguments{
\item{initMeta}{}
}
\value{
a matrix with rows for timesteps and columns for each individual in
the local community. The cells of the matrix are the species IDs. NOTE: it's
recommended that you don't work directly with this output but rather pass it
to \code{getAbund}
}
\description{
Simulate a local community under the Unified Neutral Theory of
Biodiversity.
}
