% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{untbSim}
\alias{untbSim}
\title{Unified Neutral Theory of Biodiversity Simulation}
\usage{
untbSim(
  J,
  Sm,
  Jm,
  nu,
  m,
  niter,
  niterTimestep = floor(niter/10),
  initLocal = NULL,
  initMeta = NULL
)
}
\arguments{
\item{J}{number of individuals in the local community}

\item{Sm}{number of species in the metacommunity}

\item{Jm}{number of individuals in the metacommunity}

\item{nu}{speciation probability (must be between 0 and 1)}

\item{m}{immigration probability (must be between 0 and 1)}

\item{niter}{number of iterations to run}

\item{niterTimestep}{number of iterations to save output on}

\item{initLocal}{initial state of the local community; if \code{NULL} (the
default) it will be randomly drawn from metacommunity}

\item{initMeta}{initial state of the meta community; if \code{NULL} (the
default) it will be simulated as a log series parameterized by \code{Sm} and
\code{Jm}}
}
\value{
a matrix with rows for timesteps and columns for each individual in
the local community. The cells of the matrix are the species IDs. NOTE: it's
recommended that you don't work directly with this output but rather pass it
to \code{getAbund}
}
\description{
Simulate a local community under the Unified Neutral Theory of
Biodiversity.
}
